fix nametag bug and added lookutils
